+++ src/shared/utils.rs
@@ -82,15 +82,15 @@ impl<const D: usize> WorldCoords for Point<i64, D> {
     }
 }
 
-impl<const D: usize> WorldCoords for Point<f32, D> {
+impl<const D: usize> WorldCoords for Point<f64, D> {
     type Point<T: Scalar> = Point<T, D>;
 
     fn chunk_coords(self) -> Self::Point<i32> {
-        self.map(|c| (c / Chunk::DIM as f32).floor() as i32)
+        self.map(|c| (c / Chunk::DIM as f64).floor() as i32)
     }
 
     fn block_coords(self) -> Self::Point<u8> {
-        self.map(|c| c.rem_euclid(Chunk::DIM as f32) as u8)
+        self.map(|c| c.rem_euclid(Chunk::DIM as f64) as u8)
     }
 }
